♻️ Refactor Input UI component
