+++ client/app/(user)/todo/newTodo.tsx
@@ -5,7 +5,7 @@ import { useTheme } from '@/providers/themeProvider';
 import { View } from '@/components/ui/View';
 import { Text } from '@/components/ui/Text';
 import { Button } from '@/components/ui/Button/Button';
-import { TextInput } from '@/components/ui/Input';
+import { TextInput } from '@/components/ui/Input/Input';
 
 export default function NewTodoTab() {
   const [task, setTask] = useState('');
