+++ client/app/(user)/ui-demo.tsx
@@ -6,7 +6,7 @@ import { useTheme } from '@/providers/themeProvider';
 import { View } from '@/components/ui/View';
 import { Text } from '@/components/ui/Text';
 import { Button } from '@/components/ui/Button/Button';
-import { TextInput } from '@/components/ui/Input';
+import { TextInput } from '@/components/ui/Input/Input';
 import { Card } from '@/components/ui/Card/Card';
 import { Stack, Inline, Section } from '@/components/ui/Layout';
 import { Header } from '@/components/ui/Header/Header';
