+++ client/components/ui/index.ts
@@ -2,7 +2,7 @@
 export * from './Text';
 export * from './View';
 export * from './Button/Button';
-export * from './Input';
+export * from './Input/Input';
 export * from './Card/Card';
 export * from './Layout';
 export * from './Header/Header';
